+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.0
+* Added "Pies Color" formatting options
+* Added "Show Grid", "Show Grid Values", "Show Staright Lines", "Text Color", "Text Size" options for Outer line
+* Removed stroke for pies 
+
 ## 1.0.1
 * Upgraded to API 1.7.0
 * Added language resource files